// src/types/Owner.ts
export interface OwnerImage {
    id: string;
    url: string;
    filename: string;
    size: number;
    type: string;
    width: number;
    height: number;
    thumbnails: {
      small: { url: string; width: number; height: number };
      large: { url: string; width: number; height: number };
      full: { url: string; width: number; height: number };
    };
  }
  
  export interface Owner {
    id: string;
    createdTime: string;
    name: string;
    phone: string;
    email: string;
    rut: string;
    images: OwnerImage[];
    cabin: string[];
    status: boolean;
  }
  